backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
python-telegram-bot==20.7
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
python-multipart==0.0.6
sqladmin==0.16.0
itsdangerous==2.1.2
pytoniq==0.1.43
pytoniq-core==0.1.45
redis==5.0.1
